requirements.txt
gunicorn
django-heroku
Pillow
spyrk
configparser
django-datatables-view
Faker
pandas
geojson
scipy